+++ auth_api/routers/forms.py
@@ -11,12 +11,15 @@ from ..auth import (
     get_current_active_user_opt,
     generate_ukey,
     generate_SaltNPepper,
+    decode_reset_password_token,
+    reset_password_token
 )
 from ..deps import (
     get_db,
     logger,
     oj_decode,
 )
+from datetime import timedelta
 from ..lib_mail import notify
 
 from auth_api.repositories.users import get_user_repo
@@ -55,7 +58,6 @@ async def new_user(
         ukey=generate_ukey(),
         salt=salt,
     )
-    print(UC)
 
     res = await repo.registerNewUser(user=UC)
 
@@ -64,4 +66,36 @@ async def new_user(
     return {
         'username': res['username'],
         'userid': res['userid']
+    }
+
+
+@router.post("/f/lost_password", tags=["users"])
+async def lost_password(
+        email: str = Form(...),
+        repo=Depends(get_user_repo)):
+    user = await repo.getByMail(email)
+    token = reset_password_token(user, timedelta(hours=2))
+    await notify(email, 'lost_password', {
+        'fname': user.fname,
+        'lname': user.lname,
+        'token': token
+    })
+    return {'result': True}
+
+
+@router.post("/f/reset_password", tags=["users"])
+async def lost_password(
+        token: str = Form(...),
+        password: str = Form(...),
+        repo=Depends(get_user_repo)):
+    userid = decode_reset_password_token(token)
+    salt, phash = generate_SaltNPepper(password)
+    user = await repo.getById(userid)
+    user.password = phash
+    user.salt = salt
+    await repo.updateUserPassword(user=user)
+
+    return {
+        'username': user.email,
+        'userid': user.userid
     }